# docs/api/openapi.json
{
  "components": {
    "headers": {
      "IdempotentReplay": {
        "description": "Present (value `true`) when this response was replayed from a stored `Idempotency-Key`.",
        "schema": {
          "const": "true",
          "type": "string"
        }
      },
...
    },
...
    "responses": {
...
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`FORBIDDEN` — the key lacks the scope, the request IP is outside the key's allow-list (E-70), a public key was used from an Origin outside its domain list or for anything but `lead_callback`, or the account is suspended/uninstalled."
      },
      "Internal": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`INTERNAL` — unexpected error; quote `request_id` to support. Nothing internal is disclosed."
      },
...
      "RateLimited": {
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RateLimitError"
                },
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        },
        "description": "`RATE_LIMITED`. The per-minute limiter answers with the flat `RateLimitError` body; the per-key daily intent cap (E-70) answers with the standard `Error` envelope and `Retry-After: 3600`.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        }
      },
      "Unauthenticated": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`UNAUTHENTICATED` — missing, malformed, unknown or revoked API key. Keys look like `nrd_live_…`, `nrd_test_…` or `nrd_pk_…`."
      },
...
    },
...
    "securitySchemes": {
      "publicSiteKey": {
        "bearerFormat": "nrd_pk_…",
        "description": "Public site key for browsers: `POST /v1/intents` with `use_case: lead_callback` only, from an allow-listed Origin, 10 requests/minute per IP.",
        "scheme": "bearer",
        "type": "http"
      },
      "secretKey": {
        "bearerFormat": "nrd_live_… / nrd_test_…",
        "description": "Secret API key. The scope names listed on each operation are what the key must carry.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Naaradh",
      "url": "https://naaradh.com"
    },
...
    "summary": "Two-way AI voice agent for commerce — outbound call intents and the inbound support line.",
    "title": "Naaradh API",
    "version": "0.0.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/inbound-profiles": {
      "get": {
        "description": "Every profile (draft, active, disabled) with its full configuration, most recently updated first.\n\nScope: `support:read`.",
        "operationId": "listInboundProfiles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/ProfileView"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Profiles."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "support:read"
            ]
          }
        ],
        "summary": "List inbound profiles",
        "tags": [
          "Support line"
        ]
      },
      "post": {
        "description": "Creates a profile in `draft`. The greeting must carry the AI and recording disclosure, the tools must be known, and fact limits apply — the same validator the voice runtime trusts. Activate it with POST /v1/inbound-profiles/{id}/activate once a number is attached.\n\nScope: `support:write`.",
        "operationId": "createInboundProfile",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "agent_cancel_enabled": {
                    "default": false,
                    "description": "Lets the agent cancel an unshipped COD order after the two-step confirmation (E-84). Off by default; a ticket is raised instead.",
                    "type": "boolean"
                  },
                  "business_hours": {
                    "additionalProperties": false,
                    "properties": {
                      "close": {
                        "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                        "type": "string"
                      },
                      "days": {
                        "items": {
                          "maximum": 7,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "minItems": 1,
                        "type": "array"
                      },
                      "open": {
                        "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                        "type": "string"
                      },
                      "zone": {
                        "minLength": 1,
                        "type": "string"
                      }
                    },
                    "required": [
                      "zone",
                      "days",
                      "open",
                      "close"
                    ],
                    "type": "object"
                  },
                  "closed_message": {
                    "description": "Spoken outside business hours and when the line cannot take the call.",
                    "type": "string"
                  },
                  "fallback_forward": {
                    "anyOf": [
                      {
                        "additionalProperties": false,
                        "properties": {
                          "phone": {
                            "maxLength": 32,
                            "minLength": 5,
                            "type": "string"
                          },
                          "phone_region": {
                            "default": "IN",
                            "maxLength": 2,
                            "minLength": 2,
                            "type": "string"
                          }
                        },
                        "required": [
                          "phone"
                        ],
                        "type": "object"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "description": "Staff number the line forwards to when the agent cannot answer (E-92). Encrypted with the staff key; the API can store it but never read it back — responses show a masked form."
                  },
                  "greeting": {
                    "description": "First utterance of every call. Must include the AI disclosure and recording disclosure in the profile's locale (invariant 7) or the request is a 422.",
                    "type": "string"
                  },
                  "locale": {
                    "pattern": "^[a-z]{2}-[A-Z]{2}$",
                    "type": "string"
                  },
                  "max_calls_per_caller_hour": {
                    "default": 6,
                    "maximum": 60,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "max_concurrent": {
                    "default": 2,
                    "maximum": 100,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "max_duration_sec": {
                    "default": 600,
                    "maximum": 1800,
                    "minimum": 60,
                    "type": "integer"
                  },
                  "monthly_minute_cap": {
                    "anyOf": [
                      {
                        "minimum": 1,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "description": "Inbound minutes per month before the line falls back to `fallback_forward`/the closed message."
                  },
                  "name": {
                    "maxLength": 80,
                    "minLength": 1,
                    "type": "string"
                  },
                  "persona": {
                    "default": null,
                    "description": "Tone/persona hint for the agent (max 300 characters after sanitising).",
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "pinned_facts": {
                    "default": [],
                    "description": "Short facts always available to the agent (each max 200 characters).",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "tools_enabled": {
                    "description": "Tools the agent may use on this line: `lookup_orders`, `verify_caller`, `search_knowledge`, `confirm_order`, `request_cancellation`, `request_address_change`, `create_ticket`, `transfer_to_human`, `register_opt_out`. Every tool is authorised server-side against caller identity and your settings regardless of what the model asks (invariant 18).",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1,
                    "type": "array"
                  },
                  "transfer_target_id": {
                    "default": null,
                    "description": "Default transfer target (must be one of yours).",
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "voice_id": {
                    "anyOf": [
                      {
                        "maxLength": 100,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null
                  }
                },
                "required": [
                  "name",
                  "locale",
                  "greeting",
                  "tools_enabled",
                  "closed_message",
                  "business_hours"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "description": "26-character ULID with a type prefix",
                      "type": "string"
                    },
                    "status": {
                      "const": "draft",
                      "type": "string"
                    },
                    "version": {
                      "const": 1,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "version"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created.",
            "headers": {
              "Idempotent-Replay": {
                "$ref": "#/components/headers/IdempotentReplay"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "support:write"
            ]
          }
        ],
        "summary": "Create an inbound profile",
        "tags": [
          "Support line"
        ]
      }
    },
    "/v1/inbound-profiles/{id}": {
      "put": {
        "description": "Full replacement of the configuration (send every field); the database stamps a new `version`. A null `fallback_forward` clears the stored number. Validation is the same as on create.\n\nScope: `support:write`.",
        "operationId": "replaceInboundProfile",
        "parameters": [
          {
            "description": "Profile id (`ipr_…`).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "agent_cancel_enabled": {
                    "default": false,
                    "description": "Lets the agent cancel an unshipped COD order after the two-step confirmation (E-84). Off by default; a ticket is raised instead.",
                    "type": "boolean"
                  },
                  "business_hours": {
                    "additionalProperties": false,
                    "properties": {
                      "close": {
                        "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                        "type": "string"
                      },
                      "days": {
                        "items": {
                          "maximum": 7,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "minItems": 1,
                        "type": "array"
                      },
                      "open": {
                        "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                        "type": "string"
                      },
                      "zone": {
                        "minLength": 1,
                        "type": "string"
                      }
                    },
                    "required": [
                      "zone",
                      "days",
                      "open",
                      "close"
                    ],
                    "type": "object"
                  },
                  "closed_message": {
                    "description": "Spoken outside business hours and when the line cannot take the call.",
                    "type": "string"
                  },
                  "fallback_forward": {
                    "anyOf": [
                      {
                        "additionalProperties": false,
                        "properties": {
                          "phone": {
                            "maxLength": 32,
                            "minLength": 5,
                            "type": "string"
                          },
                          "phone_region": {
                            "default": "IN",
                            "maxLength": 2,
                            "minLength": 2,
                            "type": "string"
                          }
                        },
                        "required": [
                          "phone"
                        ],
                        "type": "object"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "description": "Staff number the line forwards to when the agent cannot answer (E-92). Encrypted with the staff key; the API can store it but never read it back — responses show a masked form."
                  },
                  "greeting": {
                    "description": "First utterance of every call. Must include the AI disclosure and recording disclosure in the profile's locale (invariant 7) or the request is a 422.",
                    "type": "string"
                  },
                  "locale": {
                    "pattern": "^[a-z]{2}-[A-Z]{2}$",
                    "type": "string"
                  },
                  "max_calls_per_caller_hour": {
                    "default": 6,
                    "maximum": 60,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "max_concurrent": {
                    "default": 2,
                    "maximum": 100,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "max_duration_sec": {
                    "default": 600,
                    "maximum": 1800,
                    "minimum": 60,
                    "type": "integer"
                  },
                  "monthly_minute_cap": {
                    "anyOf": [
                      {
                        "minimum": 1,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null,
                    "description": "Inbound minutes per month before the line falls back to `fallback_forward`/the closed message."
                  },
                  "name": {
                    "maxLength": 80,
                    "minLength": 1,
                    "type": "string"
                  },
                  "persona": {
                    "default": null,
                    "description": "Tone/persona hint for the agent (max 300 characters after sanitising).",
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "pinned_facts": {
                    "default": [],
                    "description": "Short facts always available to the agent (each max 200 characters).",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "tools_enabled": {
                    "description": "Tools the agent may use on this line: `lookup_orders`, `verify_caller`, `search_knowledge`, `confirm_order`, `request_cancellation`, `request_address_change`, `create_ticket`, `transfer_to_human`, `register_opt_out`. Every tool is authorised server-side against caller identity and your settings regardless of what the model asks (invariant 18).",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1,
                    "type": "array"
                  },
                  "transfer_target_id": {
                    "default": null,
                    "description": "Default transfer target (must be one of yours).",
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "voice_id": {
                    "anyOf": [
                      {
                        "maxLength": 100,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null
                  }
                },
                "required": [
                  "name",
                  "locale",
                  "greeting",
                  "tools_enabled",
                  "closed_message",
                  "business_hours"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            },
            "description": "The updated profile."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "support:write"
            ]
          }
        ],
        "summary": "Replace an inbound profile",
        "tags": [
          "Support line"
        ]
      }
    },
...
    "/v1/tickets": {
      "get": {
        "description": "What the agent could not do itself — callbacks, address changes, cancellations it was not allowed to execute, refunds, complaints — up to 200, highest priority first then newest. Every ticket is also announced by the `ticket.created` webhook.\n\nScope: `tickets:read`.",
        "operationId": "listTickets",
        "parameters": [
          {
            "description": "Filter by status; omitted returns every status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "open",
                "in_progress",
                "resolved"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/Ticket"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Tickets."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "tickets:read"
            ]
          }
        ],
        "summary": "List tickets",
        "tags": [
          "Support line"
        ]
      }
    },
...
  },
  "security": [
    {
      "secretKey": []
    }
  ],
  "servers": [
    {
      "description": "Production",
      "url": "https://api.naaradh.com"
    },
    {
      "description": "Local development (apps/api default PORT)",
      "url": "http://localhost:3001"
    }
  ],
...
}